+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 apidoc extension Change Log
 
 - Enh #185: Use HTTPS for www.php.net links (kamarton)
 - Bug #187: Prevent getter/setter methods from affecting class-defined property docs (brandonkelly)
+- Enh #137: @since tags are now propagated to inherited methods/properties in the same package (brandonkelly)
 - Bug #145: Fixed broken API links on property/method docs that were pulled in with @inheritdoc (brandonkelly)